Move spawn errors into executor.rs
